Add separate .eslintrcjs file for storybook
- Fix imports for stories inside their own directory
- Add lint:stories script
